refactor: Extract `TriggerErrorInfo` description as a component
